Split router into its separate file
